Refactoring (#58)
