+++ snippets/math/prime.py
